Split TOU into provider/controller
